requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.0
idna==2.8
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
whitenoise==4.1.4
